Sauvegarde du 30/06/2024

+++ app/Http/Controllers/TypeDocumentController.php
@@ -2,15 +2,15 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\TypeDocumentRequest;
 use App\Models\TypeDocument;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Request;
 use Illuminate\Support\Facades\Storage;
 
 class TypeDocumentController extends Controller
 {
-    
-    
+
+
     public function index()
     {
         $typeDocuments = TypeDocument::all();
@@ -19,51 +19,56 @@ class TypeDocumentController extends Controller
     }
 
     //Fonction Post pour ajouter un guide dans la base de données
-    public function create(Request $request)
+    public function create(TypeDocumentRequest $request)
     {
-        $request->validate([
-            'nom_fichier' => 'required',
-            'fichier_scanner' => 'required',
-        ]);
-        
-        $ancienFichier = "";
-        //On va remplacer le repertoire "storage" par "public" concernant
-        //l'ancien fichier du document
-        $ancienFichier = str_replace('storage/', 'public/', $ancienFichier);
-        
-        if ($ancienFichier && Storage::exists($ancienFichier)) {
-            Storage::delete($ancienFichier);
+        $request->validated();
+
+        //Générer l'empreinte numérique avec la fonction MD5
+        $empreinte = md5($request->hasFile('fichier_scanner'));
+
+        //Récupérer une empreinte existant
+        $existe = TypeDocument::where('empreinte_fichier', $empreinte)->exists();
+        if ($existe) {
+            // L'empreinte du fichier existe déjà dans la base de données -> on affiche un warning.
+            return redirect()->route('type-document.index')->with('warning', 'Le document existe déjà dans la base de données.');
+        } else {
+            $ancienFichier = "";
+            //On va remplacer le repertoire "storage" par "public" concernant
+            //l'ancien fichier du document
+            $ancienFichier = str_replace('storage/', 'public/', $ancienFichier);
+
+            if ($ancienFichier && Storage::exists($ancienFichier)) {
+                Storage::delete($ancienFichier);
+            }
+
+            //Stocké l'image dans la variable $path dans
+            //le répertoire /storage/public/upload
+            //et créer un lien du repertoire storage vers public
+            $path = $request->file('fichier_scanner')->store('public/upload');
+            //Rétirer le repertoire public avant de mettre dans la base de données
+            $replace_path = str_replace('public', '', $path);
+            // L'empreinte du fichier n'existe pas dans la base de données
+            $typeDocuments = new TypeDocument();
+            $typeDocuments->nom_fichier = $request->nom_fichier;
+            //ajouter le repertoire "storage" pour l'insertion de la base de données
+            $typeDocuments->fichier_scanner = "storage" . $replace_path;
+            $typeDocuments->empreinte_fichier = $empreinte;
+            $typeDocuments->save();
+            //Après avoir enregistrer, faire la rediretion
+            return redirect()->route('type-document.index')->with('status', 'Le document a bien été enregistré.');
         }
-
-        //Stocké l'image dans la variable $path dans
-        //le répertoire /storage/public/upload
-        //et créer un lien du repertoire storage vers public
-        $path = $request->file('fichier_scanner')->store('public/upload');
-        //Rétirer le repertoire public avant de mettre dans la base de données
-        $replace_path = str_replace('public', '', $path);
-
-        $typeDocuments = new TypeDocument();
-        $typeDocuments->nom_fichier = $request->nom_fichier;
-        //ajouter le repertoire "storage" pour l'insertion de la base de données
-        $typeDocuments->fichier_scanner = "storage". $replace_path;  
-        $typeDocuments->empreinte_fichier = Hash::make($path);          
-        $typeDocuments->save();
-
-        //Après avoir enregistrer, faire la rediretion
-        return redirect()->route('type-document.index')->with('status', 'Le document a bien été enregistré.');
-
     }
 
     //Delete TypeDocument
     public function delete($id)
     {
         $typeDocument = TypeDocument::find($id); //Récupéré le document
-        
+
         $ancienFichier = $typeDocument->fichier_scanner;
         //On va remplacer le repertoire "storage" par "public" concernant
         //l'ancien fichier du document
         $ancienFichier = str_replace('storage/', 'public/', $ancienFichier);
-        
+
         if ($ancienFichier && Storage::exists($ancienFichier)) {
             Storage::delete($ancienFichier);
         }
@@ -72,6 +77,4 @@ class TypeDocumentController extends Controller
         //Après avoir supprimer, faire la rediretion
         return redirect()->route('type-document.index')->with('status', 'Le document a bien été supprimé.');
     }
-    
-
 }
